moved common resources to Shared Library
