+++ src/JoeScan.LogScanner.Desktop/App.xaml
@@ -12,12 +12,12 @@
                     <desktop:AppBootstrapper x:Key="Bootstrapper" />
                 </ResourceDictionary>
                 <ResourceDictionary>
-                    <FontFamily x:Key="InterUI">pack://application:,,,/Fonts/#Inter Medium</FontFamily>
-                    <FontFamily x:Key="InterThin">pack://application:,,,/Fonts/#Inter Thin</FontFamily>
-                    <FontFamily x:Key="InterLight">pack://application:,,,/Fonts/#Inter Light</FontFamily>
-                    <FontFamily x:Key="InterUIBlack">pack://application:,,,/Fonts/#Inter Black</FontFamily>
-                    <FontFamily x:Key="InterUIBold">pack://application:,,,/Fonts/#Inter Bold</FontFamily>
-                    <FontFamily x:Key="MaterialDesignIcons">pack://application:,,,/Fonts/#Material Design Icons</FontFamily>
+                    <FontFamily x:Key="InterUI">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Medium</FontFamily>
+                    <FontFamily x:Key="InterThin">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Thin</FontFamily>
+                    <FontFamily x:Key="InterLight">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Light</FontFamily>
+                    <FontFamily x:Key="InterUIBlack">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Black</FontFamily>
+                    <FontFamily x:Key="InterUIBold">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Bold</FontFamily>
+                    <FontFamily x:Key="MaterialDesignIcons">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Material Design Icons</FontFamily>
                 </ResourceDictionary>
                 <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/ColorSchemes/dark.xaml" />
                 <!-- <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/ColorSchemes/light.xaml" /> -->
