--- src/JoeScan.LogScanner.Desktop/LogProperties/LogPropertiesView.xaml
@@ -1,43 +1,0 @@
-﻿<UserControl
-    x:Class="JoeScan.LogScanner.Desktop.LogProperties.LogPropertiesView"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:dd="urn:gong-wpf-dragdrop"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    d:DesignHeight="450"
-    d:DesignWidth="800"
-    mc:Ignorable="d">
-    <Grid>
-        <ListBox
-            x:Name="Items"
-            dd:DragDrop.IsDragSource="True"
-            dd:DragDrop.IsDropTarget="True">
-            <ListBox.Style>
-                <Style BasedOn="{StaticResource {x:Type ListBox}}" TargetType="ListBox">
-                    <Setter Property="ItemsPanel">
-                        <Setter.Value>
-                            <ItemsPanelTemplate>
-                                <WrapPanel />
-                            </ItemsPanelTemplate>
-                        </Setter.Value>
-                    </Setter>
-                    <Setter Property="ScrollViewer.HorizontalScrollBarVisibility" Value="Disabled" />
-                </Style>
-            </ListBox.Style>
-            <ListBox.ItemContainerStyle>
-                <Style BasedOn="{StaticResource {x:Type ListBoxItem}}" TargetType="{x:Type ListBoxItem}">
-                    <Setter Property="Focusable" Value="False" />
-                    <Setter Property="Padding" Value="0" />
-                    <Setter Property="Margin" Value="2" />
-                    <Style.Triggers>
-                        <Trigger Property="IsSelected" Value="True">
-                            <Setter Property="Background" Value="transparent" />
-                        </Trigger>
-                    </Style.Triggers>
-                </Style>
-            </ListBox.ItemContainerStyle>
-        </ListBox>
-
-    </Grid>
-</UserControl>
