--- src/JoeScan.LogScanner.Desktop/LogProperties/LogPropertiesView.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace JoeScan.LogScanner.Desktop.LogProperties
-{
-    /// <summary>
-    /// Interaction logic for LogDataView.xaml
-    /// </summary>
-    public partial class LogPropertiesView : UserControl
-    {
-        public LogPropertiesView()
-        {
-            InitializeComponent();
-        }
-    }
-}
