--- src/JoeScan.LogScanner.Desktop/LogProperties/LogPropertyItemView.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace JoeScan.LogScanner.Desktop.LogProperties
-{
-    /// <summary>
-    /// Interaction logic for LogPropertyItemView.xaml
-    /// </summary>
-    public partial class LogPropertyItemView : UserControl
-    {
-        public LogPropertyItemView()
-        {
-            InitializeComponent();
-        }
-    }
-}
