+++ src/JoeScan.LogScanner.Desktop/Shell/ShellViewModel.cs
@@ -4,11 +4,11 @@ using JoeScan.LogScanner.Core.Models;
 using JoeScan.LogScanner.Desktop.Config;
 using JoeScan.LogScanner.Desktop.LiveProfiles;
 using JoeScan.LogScanner.Desktop.LogHistory;
-using JoeScan.LogScanner.Desktop.LogProperties;
 using JoeScan.LogScanner.Desktop.StatusBar;
 using JoeScan.LogScanner.Desktop.Toolbar;
 using JoeScan.LogScanner.Desktop.TopAndSide;
 using JoeScan.LogScanner.Shared.Live3D;
+using JoeScan.LogScanner.Shared.LogProperties;
 using JoeScan.LogScanner.Shared.Notifier;
 using NLog;
 using System.Threading;
@@ -60,6 +60,7 @@ public class ShellViewModel : Screen
         LogHistory = logHistory;
         Live3D = live3D;
         LogProperties = logProperties;
+        LogProperties.SetDisplayUnits(config.Units);
         Engine = engine;
         Logger = logger;
         Config = config;
