+++ src/JoeScan.LogScanner.LogReview/App.xaml
@@ -11,7 +11,14 @@
                 <ResourceDictionary>
                     <local:AppBootstrapper x:Key="Bootstrapper" />
                 </ResourceDictionary>
-
+                <ResourceDictionary>
+                    <FontFamily x:Key="InterUI">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Medium</FontFamily>
+                    <FontFamily x:Key="InterThin">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Thin</FontFamily>
+                    <FontFamily x:Key="InterLight">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Light</FontFamily>
+                    <FontFamily x:Key="InterUIBlack">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Black</FontFamily>
+                    <FontFamily x:Key="InterUIBold">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Inter Bold</FontFamily>
+                    <FontFamily x:Key="MaterialDesignIcons">pack://application:,,,/JoeScan.LogScanner.Shared;component/Fonts/#Material Design Icons</FontFamily>
+                </ResourceDictionary>
                 <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/ColorSchemes/dark.xaml" />
                 <!-- <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/ColorSchemes/light.xaml" /> -->
                 <ResourceDictionary Source="pack://application:,,,/AdonisUI.ClassicTheme;component/Resources.xaml" />
