+++ src/JoeScan.LogScanner.LogReview/AppBootstrapper.cs
@@ -13,6 +13,7 @@ using JoeScan.LogScanner.LogReview.Models;
 using JoeScan.LogScanner.LogReview.SectionTable;
 using JoeScan.LogScanner.LogReview.Settings;
 using JoeScan.LogScanner.Shared.Live3D;
+using JoeScan.LogScanner.Shared.LogProperties;
 using MvvmDialogs;
 using JoeScan.LogScanner.Shared.Notifier;
 
@@ -43,6 +44,9 @@ public class AppBootstrapper : AutofacBootstrapper
         // re-register the Live View so it uses the constructor that does not require a whole log 
         // engine, so we don't pull all the other crap in that is not needed here
         builder.RegisterType<Live3DViewModel>().UsingConstructor();
+        // re-register the LogProperties view so it uses the constructor that does not require a whole log 
+        // engine, so we don't pull all the other crap in that is not needed here
+        builder.RegisterType<LogPropertiesViewModel>().UsingConstructor();
     }
 }
