+++ src/JoeScan.LogScanner.LogReview/Settings/ILogReviewSettings.cs
@@ -1,6 +1,11 @@
-﻿namespace JoeScan.LogScanner.LogReview.Settings;
+﻿using Config.Net;
+using JoeScan.LogScanner.Shared.Enums;
+
+namespace JoeScan.LogScanner.LogReview.Settings;
 
 public interface ILogReviewSettings
 {
     string FileBrowserLastFolder { get; set; }
+    [Option(DefaultValue = "Inches")]
+    DisplayUnits Units { get; set; }
 }
