+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-Black.ttf
