+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-Bold.ttf
