+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-ExtraBold.ttf
