+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-ExtraLight.ttf
