+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-Light.ttf
