+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-Medium.ttf
