+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-Regular.ttf
