+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-SemiBold.ttf
