+++ src/JoeScan.LogScanner.Shared/Fonts/Inter-Thin.ttf
