+++ src/JoeScan.LogScanner.Shared/Fonts/materialdesignicons-webfont.ttf
