+++ src/JoeScan.LogScanner.Shared/LogProperties/LogPropertiesView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace JoeScan.LogScanner.Shared.LogProperties
+{
+    /// <summary>
+    /// Interaction logic for LogDataView.xaml
+    /// </summary>
+    public partial class LogPropertiesView : UserControl
+    {
+        public LogPropertiesView()
+        {
+            InitializeComponent();
+        }
+    }
+}
