+++ src/JoeScan.LogScanner.Shared/LogProperties/LogPropertyItemView.xaml
@@ -1,0 +1,61 @@
+﻿<UserControl
+    x:Class="JoeScan.LogScanner.Shared.LogProperties.LogPropertyItemView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:adonisUi="clr-namespace:AdonisUI;assembly=AdonisUI"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
+    mc:Ignorable="d">
+    <Border
+        MinWidth="180"
+        MaxWidth="180"
+        Padding="3"
+        Background="{DynamicResource {x:Static adonisUi:Brushes.Layer2BackgroundBrush}}"
+        BorderThickness="1"
+        CornerRadius="3">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="*" />
+                <RowDefinition Height="Auto" />
+            </Grid.RowDefinitions>
+            <DockPanel Grid.Row="0">
+                <!-- <Button DockPanel.Dock="Right" Style="{DynamicResource {x:Static adonisUi:Styles.ToolbarButton}}"> -->
+                <!--  <TextBlock  -->
+                <!--  VerticalAlignment="Center"  -->
+                <!--  FontFamily="{StaticResource MaterialDesignIcons}"  -->
+                <!--  FontSize="16">  -->
+                <!--  &#xF01D8;  -->
+                <!--     </TextBlock> -->
+                <!-- </Button> -->
+                <TextBlock
+                    Padding="2"
+                    DockPanel.Dock="Left"
+                    FontFamily="{StaticResource InterLight}"
+                    Text="{Binding PropertyName}" />
+            </DockPanel>
+            <DockPanel Grid.Row="1">
+                <TextBlock
+                    Padding="2"
+                    VerticalAlignment="Bottom"
+                    DockPanel.Dock="Right"
+                    FontFamily="{StaticResource InterThin}"
+                    Text="{Binding UnitString}"
+                    TextAlignment="Right" />
+
+                <TextBlock
+                    Padding="2"
+                    VerticalAlignment="Center"
+                    DockPanel.Dock="Left"
+                    FontFamily="{StaticResource InterUIBold}"
+                    FontSize="22"
+                    Foreground="{DynamicResource {x:Static adonisUi:Brushes.AccentBrush}}"
+                    Text="{Binding PropertyValue}"
+                    TextAlignment="Right" />
+            </DockPanel>
+        </Grid>
+    </Border>
+</UserControl>
