+++ src/JoeScan.LogScanner.Shared/LogProperties/LogPropertyItemView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace JoeScan.LogScanner.Shared.LogProperties
+{
+    /// <summary>
+    /// Interaction logic for LogPropertyItemView.xaml
+    /// </summary>
+    public partial class LogPropertyItemView : UserControl
+    {
+        public LogPropertyItemView()
+        {
+            InitializeComponent();
+        }
+    }
+}
